GRIDEDIT-265: change NextCircularForwardIndex and NextCircularBackwardIndex to use size_t parameters and return size_t

+++ include/MeshKernel/Mesh.hpp
@@ -373,7 +373,7 @@ namespace meshkernel
 
         // faces
         std::vector<std::vector<size_t>> m_facesNodes; ///< The nodes composing the faces, in ccw order (netcell%Nod)
-        std::vector<int> m_numFacesNodes;              ///< The number of nodes composing the face (netcell%N)
+        std::vector<size_t> m_numFacesNodes;           ///< The number of nodes composing the face (netcell%N)
         std::vector<std::vector<size_t>> m_facesEdges; ///< The edge indices composing the face (netcell%lin)
         std::vector<Point> m_facesCircumcenters;       ///< The face circumcenters the face circumcenter (xz, yz)
         std::vector<Point> m_facesMassCenters;         ///< The faces centers of mass (xzw, yzw)
