+++ include/MeshKernel/MeshRefinement.hpp
@@ -150,9 +150,9 @@ namespace meshkernel
         /// The sample node RTree
         SpatialTrees::RTree m_samplesRTree;
 
-        std::vector<int> m_faceMask;     ///< Compute face without hanging nodes (1), refine face with hanging nodes (2), do not refine cell at all (0) or refine face outside polygon (-2)
-        std::vector<int> m_edgeMask;     ///< If 0, edge is not split
-        std::vector<int> m_brotherEdges; ///< The index of the brother edge for each edge
+        std::vector<int> m_faceMask;        ///< Compute face without hanging nodes (1), refine face with hanging nodes (2), do not refine cell at all (0) or refine face outside polygon (-2)
+        std::vector<int> m_edgeMask;        ///< If 0, edge is not split
+        std::vector<size_t> m_brotherEdges; ///< The index of the brother edge for each edge
 
         /// Local caches
         std::vector<bool> m_isHangingNodeCache;       ///< Cache for maintaining if node is hanging
